move tensor tree related methods to separate package
