+++ core/src/main/scala/dimwit/package.scala
@@ -84,7 +84,9 @@ package object dimwit:
   // Export devices
   export dimwit.hardware.Device
   // Export automatic differentiation
-  export dimwit.autodiff.{Autodiff, TensorTree, FloatTree, Grad}
+  export dimwit.autodiff.{Autodiff, Grad}
+  // Export tensor trees
+  export dimwit.tensortree.{TensorTree, TensorTreeIO, TensorTreeFormat, FloatTree}
   // Export Just-in-Time compilation
   export dimwit.jax.Jit.{jit, jitDonating, jitDonatingUnsafe}
   export dimwit.jax.EagerCleanup.eagerCleanup
